+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**[0.38.0] – 2025-XX-XX**  
+
+**Umfangreiche Änderungen durch Firmwareupdate auf 1.38.6-1**  
+- Firmwareversionen kleiner 1.36.6-1 werden nicht mehr unterstützt.  
+- Nach einem Firmwareupdate auf 1.38.6-1 muss das Kennwort  für `customer` neu gesetzt werden.  
+
 **[0.31.5] – 2025-09-02**  
 
 Neukonzipierung von FUNCTIONS/GEN24_API.py: